+++ local.js
@@ -5157,4 +5157,7 @@ function Local__submitHandler(e) {
 }
 
 local.bindRequestEvents = bindRequestEvents;
-local.unbindRequestEvents = unbindRequestEvents;})();
+local.unbindRequestEvents = unbindRequestEvents;
+
+// Turn on by default
+local.bindRequestEvents(document.body);})();